refactor: exercise card component split into resistance and cardio cards
